Disable some code regarding updating
