+++ src/utils/functionRuntimeUtils.ts
@@ -3,19 +3,19 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as opn from 'opn';
+// import * as opn from 'opn';
 import * as semver from 'semver';
-import * as vscode from 'vscode';
-import { callWithTelemetryAndErrorHandling, DialogResponses, IActionContext, parseError } from 'vscode-azureextensionui';
-import { attemptToInstallLatestFunctionRuntime, brewOrNpmInstalled } from '../commands/createNewProject/validateFuncCoreToolsInstalled';
+// import * as vscode from 'vscode';
+import { callWithTelemetryAndErrorHandling, IActionContext, parseError } from 'vscode-azureextensionui';
+// import { attemptToInstallLatestFunctionRuntime, brewOrNpmInstalled } from '../commands/createNewProject/validateFuncCoreToolsInstalled';
 import { isWindows, ProjectRuntime } from '../constants';
 import { ext } from '../extensionVariables';
-import { localize } from '../localize';
-import { getFuncExtensionSetting, updateGlobalSetting } from '../ProjectSettings';
+// import { localize } from '../localize';
+import { getFuncExtensionSetting } from '../ProjectSettings';
+import { cliFeedJsonResponse, getCliFeedJson } from '../utils/getCliFeedJson';
 import { cpUtils } from './cpUtils';
 
 export namespace functionRuntimeUtils {
-    const runtimePackage: string = 'azure-functions-core-tools';
     enum FunctionRuntimeTag {
         latest = 1,
         core = 2
@@ -39,36 +39,39 @@ export namespace functionRuntimeUtils {
                     if (newestVersion === null) {
                         return;
                     }
-                    if (semver.gt(newestVersion, localVersion)) {
-                        const canUpdate: boolean = await brewOrNpmInstalled();
-                        const message: string = canUpdate ? localize(
-                            'azFunc.outdatedFunctionRuntimeUpdate',
-                            'Your version of the Azure Functions Core Tools ({0}) does not match the latest ({1}). Would you like to update now?',
-                            localVersion,
-                            newestVersion
-                        ) : localize(
-                            'azFunc.outdatedFunctionRuntimeSeeMore',
-                            'Your version of the Azure Functions Core Tools ({0}) does not match the latest ({1}). Please update for the best experience.',
-                            localVersion,
-                            newestVersion);
 
-                        const result: vscode.MessageItem = await ext.ui.showWarningMessage(message, canUpdate ? DialogResponses.yes : DialogResponses.learnMore, DialogResponses.dontWarnAgain);
-                        if (result === DialogResponses.learnMore) {
-                            // tslint:disable-next-line:no-unsafe-any
-                            opn('https://aka.ms/azFuncOutdated');
-                        } else if (result === DialogResponses.yes) {
-                            switch (major) {
-                                case FunctionRuntimeTag.latest:
-                                    await attemptToInstallLatestFunctionRuntime('v1');
-                                case FunctionRuntimeTag.core:
-                                    await attemptToInstallLatestFunctionRuntime('v2');
-                                default:
-                                    break;
-                            }
-                        } else if (result === DialogResponses.dontWarnAgain) {
-                            await updateGlobalSetting(settingKey, false);
-                        }
-                    }
+                    /* Code disabled untiled we can verify a consistent way to determine when a new release is available
+                    * if (semver.gt(newestVersion, localVersion)) {
+                    *     const canUpdate: boolean = await brewOrNpmInstalled();
+                    *     const message: string = canUpdate ? localize(
+                    *         'azFunc.outdatedFunctionRuntimeUpdate',
+                    *         'Your version of the Azure Functions Core Tools ({0}) does not match the latest ({1}). Would you like to update now?',
+                    *         localVersion,
+                    *         newestVersion
+                    *     ) : localize(
+                    *         'azFunc.outdatedFunctionRuntimeSeeMore',
+                    *         'Your version of the Azure Functions Core Tools ({0}) does not match the latest ({1}). Please update for the best experience.',
+                    *         localVersion,
+                    *         newestVersion);
+
+                    *     const result: vscode.MessageItem = await ext.ui.showWarningMessage(message, canUpdate ? DialogResponses.yes : DialogResponses.learnMore, DialogResponses.dontWarnAgain);
+                    *     if (result === DialogResponses.learnMore) {
+                    *         // tslint:disable-next-line:no-unsafe-any
+                    *         opn('https://aka.ms/azFuncOutdated');
+                    *     } else if (result === DialogResponses.yes) {
+                    *         switch (major) {
+                    *             case FunctionRuntimeTag.latest:
+                    *                 await attemptToInstallLatestFunctionRuntime('v1');
+                    *             case FunctionRuntimeTag.core:
+                    *                 await attemptToInstallLatestFunctionRuntime('v2');
+                    *             default:
+                    *                 break;
+                    *         }
+                    *     } else if (result === DialogResponses.dontWarnAgain) {
+                    *         await updateGlobalSetting(settingKey, false);
+                    *     }
+                    * }
+                    */
                 } catch (error) {
                     ext.outputChannel.appendLine(`Error occurred when checking the version of 'Azure Functions Core Tools': ${parseError(error).message}`);
                     throw error;
@@ -105,25 +108,23 @@ export namespace functionRuntimeUtils {
         const versionInfo: string = await cpUtils.executeCommand(undefined, undefined, 'func');
         const matchResult: RegExpMatchArray | null = versionInfo.match(/(?:.*)Azure Functions Core Tools (.*)/);
         if (matchResult !== null) {
-            const localVersion: string = matchResult[1].replace(/[()]/g, '').trim(); // remove () and whitespace
+            let localVersion: string = matchResult[1].replace(/[()]/g, '').trim(); // remove () and whitespace
+            // this is a fix for a bug currently in the Function CLI
+            if (localVersion === '220.0.0-beta.0') {
+                localVersion = '2.0.1-beta.25';
+            }
             return semver.valid(localVersion);
         }
         return null;
     }
 
     async function getNewestFunctionRuntimeVersion(major: number): Promise<string | null> {
+        const cliFeedJson: cliFeedJsonResponse = await getCliFeedJson();
         switch (major) {
             case FunctionRuntimeTag.latest:
-                return semver.valid((await cpUtils.executeCommand(undefined, undefined, 'npm', 'view', runtimePackage, 'dist-tags.latest')).trim());
+                return cliFeedJson.tags.v1.release;
             case FunctionRuntimeTag.core:
-                try {
-                    // since Mac OS and Windows may have installed through npm, try that first
-                    return semver.valid((await cpUtils.executeCommand(undefined, undefined, 'npm', 'view', runtimePackage, 'dist-tags.core')).trim());
-                } catch (error) {
-                    // Currently there is no way to check if `brew outdated` works because there are no old versions available of azure-function-core-tools
-                    return null;
-                }
-
+                return cliFeedJson.tags.v2.release;
             default:
                 return null;
         }
